+++ test/karma/lforms-data.spec.js
@@ -19,5 +19,38 @@ describe('LFormsData class', function() {
       assert(lfData.lformsVersion.length > 0);
     });
 
+    it('should set validation for answer-list items that must match the list', function() {
+      var lfData = new LForms.LFormsData({name: 'test form', items: [
+        {
+          linkId: 'options-only',
+          question: 'Options only autocomplete',
+          dataType: 'ST',
+          answers: [{text: 'A'}, {text: 'B'}]
+        },
+        {
+          linkId: 'options-or-string',
+          question: 'Options or string',
+          dataType: 'CODING',
+          answerConstraint: 'optionsOrString',
+          answers: [{text: 'A'}, {text: 'B'}]
+        },
+        {
+          linkId: 'options-only-radio-checkbox',
+          question: 'Options only radio checkbox',
+          dataType: 'ST',
+          displayControl: {
+            answerLayout: {
+              type: 'RADIO_CHECKBOX'
+            }
+          },
+          answers: [{text: 'A'}, {text: 'B'}]
+        }
+      ]});
+
+      assert.equal(lfData.items[0]._hasValidation, true);
+      assert.equal(lfData.items[1]._hasValidation, undefined);
+      assert.equal(lfData.items[2]._hasValidation, undefined);
+    });
+
   });
 });