+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [43.2.0] 2026-09-08
+## [44.1.0] 2026-09-08
 ### Added
 - Support the `check-box` item control on FHIR Questionnaire boolean items with
   a localized three-state checkbox for Yes, No, and Not Answered.
+
+## [44.0.0] 2026-08-11
+### Added
+- Validation of values entered in autocomplete fields, with a red border and
+  tooltip error message shown when the value does not match an option in the list.
+### Changed
+- LForms.Util.checkConstraints() is now an async function.
 
 ## [43.1.0] 2026-08-10
 ### Changed